+++ Two-Factor/app/src/main/AndroidManifest.xml
@@ -1,5 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
+<?xml version="1.0" encoding="utf-8"?><!--
   JBoss, Home of Professional Open Source
   Copyright Red Hat, Inc., and individual contributors
 
@@ -15,20 +14,21 @@
   See the License for the specific language governing permissions and
   limitations under the License.
 -->
-<manifest
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="org.jboss.aerogear.android.cookbook.twofactor"
-    xmlns:android="http://schemas.android.com/apk/res/android"
     android:versionCode="1"
     android:versionName="1.0">
 
-    <uses-sdk android:minSdkVersion="16" android:maxSdkVersion="21" />
+    <uses-sdk
+        android:maxSdkVersion="21"
+        android:minSdkVersion="16" />
 
     <application
         android:icon="@drawable/ic_launcher"
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".OTPMainActivity"
+            android:name=".ui.OTPMainActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait">
             <intent-filter>
@@ -37,15 +37,11 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".OTPLoginActivity"
+            android:name=".ui.OTPQRCodeActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".OTPQRCodeActivity"
-            android:label="@string/app_name"
-            android:screenOrientation="portrait" />
-        <activity
-            android:name=".OTPCodeActivity"
+            android:name=".ui.OTPCodeActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait" />
     </application>
