--- Two-Factor/app/src/main/res/layout/display.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              android:orientation="vertical"
-              android:layout_width="fill_parent"
-              android:layout_height="fill_parent"
-              android:gravity="center"
-              android:padding="30dp">
-
-    <ImageView
-            android:layout_width="250dp"
-            android:layout_height="80dp"
-            android:layout_gravity="center"
-            android:contentDescription="@string/aerogear_logo_content_description"
-            android:src="@drawable/aerogear_logo"/>
-
-    <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="right"
-            android:paddingBottom="30dp"
-            android:textSize="20dp"
-            android:textStyle="bold"
-            android:text="@string/project_title"/>
-
-    <TextView
-            android:id="@+id/name"
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:gravity="center_horizontal"
-            android:textSize="24dp"/>
-
-    <ProgressBar
-            android:id="@+id/progress"
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:gravity="center_horizontal"
-            android:paddingBottom="10dp"
-            style="@android:style/Widget.Holo.ProgressBar.Horizontal"/>
-
-    <TextView
-            android:id="@+id/totp"
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:gravity="center_horizontal"
-            android:textSize="20dp"/>
-
-</LinearLayout>
-
