+++ Two-Factor/app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="app_name">Two Factor</string>
-    <string name="project_title">Two Factor Auth / OTP Demo</string>
+    <string name="project_title">Two Factor</string>
     <string name="scan_qrcode">Scan QR Code</string>
     <string name="warning">Warning</string>
     <string name="error_occurred_qrcode">An error occurred while trying to read the QRCode. Do you want to try again?</string>
